atualizando banco

+++ model/Demand.py
@@ -7,10 +7,10 @@ class DemandType(orm.Model):
     name = orm.Column(orm.String(200))
 
 class Demand(orm.Model):
-    id = orm.Column(orm.Integer, primary_key=True)
+    id = orm.Column(orm.BigInteger, primary_key=True)
     name = orm.Column(orm.String(200))
     description = orm.Column(orm.String(500))
-    type_id = orm.Column(orm.BigInteger, orm.ForeignKey('demand_type.id'))
+    type_id = orm.Column(orm.Integer, orm.ForeignKey('demand_type.id'))
     creationDate = orm.Column(orm.DateTime, nullable=False, default=datetime.now())
     def __init__(self, name:str, description:str, demandType: DemandType):
         self.name = name
